+++ packages/bun-plugin-preact/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tsrx/bun-plugin-preact
 
+## 0.0.97
+
+### Patch Changes
+
+- Updated dependencies
+  [[`2d053f4`](https://github.com/tsrx-org/tsrx/commit/2d053f421f09c5c4936d8866bbeb44b924411dbb)]:
+  - @tsrx/core@0.2.2
+  - @tsrx/preact@0.2.2
+
+## 0.0.96
+
+### Patch Changes
+
+- Updated dependencies
+  [[`33d6093`](https://github.com/tsrx-org/tsrx/commit/33d60939c720d7d0bb8ab03790486a4c033dde96)]:
+  - @tsrx/core@0.2.1
+  - @tsrx/preact@0.2.1
+
 ## 0.0.95
 
 ### Patch Changes