+++ packages/rspack-plugin-solid/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tsrx/rspack-plugin-solid
 
+## 0.0.87
+
+### Patch Changes
+
+- Updated dependencies
+  [[`2d053f4`](https://github.com/tsrx-org/tsrx/commit/2d053f421f09c5c4936d8866bbeb44b924411dbb)]:
+  - @tsrx/core@0.2.2
+  - @tsrx/solid@0.2.2
+
+## 0.0.86
+
+### Patch Changes
+
+- Updated dependencies
+  [[`33d6093`](https://github.com/tsrx-org/tsrx/commit/33d60939c720d7d0bb8ab03790486a4c033dde96)]:
+  - @tsrx/core@0.2.1
+  - @tsrx/solid@0.2.1
+
 ## 0.0.85
 
 ### Patch Changes